SV-Comp Submission for 2026 (Artifact/ Replication package at: https://zenodo.org/records/17748741)

+++ symbolic-executor/src/main/java/de/uzl/its/swat/symbolic/trace/Element.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @JsonTypeInfo(use = JsonTypeInfo.Id.CLASS, include = JsonTypeInfo.As.PROPERTY, property = "@class")
 class Element implements Serializable {
     // Unique identifier for the element
-    private int iid;
+    private long iid;
 }
